[Testing] Renamed __tests__ directory to tests
