+++ lib/components/related_artists/__tests__/layout-test.js
@@ -1,7 +1,7 @@
 import 'react-native'
 import React from 'react'
 import RelatedArtists from '../'
-import { renderWithLayout } from '../../../testing_utilities/helpers'
+import { renderWithLayout } from '../../../tests/renderWithLayout'
 
 jest.mock('../../opaque_image_view.js', () => 'ImageView')
